refactor(reservation/store) : HalfHourOnly 어노테이션 수정 #85
### 변경사항
- 가게, 예약 도메인 모두 동일한 HalfHourOnly 어노테이션을 사용하도록 수정
- HalfHourOnly에 2개의 Validator를 등록
