--- src/main/java/org/example/tablenow/domain/reservation/annotation/HalfHourOnly.java
@@ -1,16 +1,0 @@
-package org.example.tablenow.domain.reservation.annotation;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.*;
-
-@Documented
-@Constraint(validatedBy = HalfHourValidator.class)
-@Target({ ElementType.FIELD })
-@Retention(RetentionPolicy.RUNTIME)
-public @interface HalfHourOnly {
-    String message() default "예약 시간은 0분 또는 30분이어야 합니다.";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
