+++ src/main/java/org/example/tablenow/domain/reservation/dto/request/ReservationUpdateRequestDto.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.tablenow.domain.reservation.annotation.HalfHourOnly;
+import org.example.tablenow.global.annotation.HalfHourOnly;
 
 import java.time.LocalDateTime;
